src/han/OOD/parkingLot/ParkingLot.java
package han.OOD.parkingLot;

/**
 * Created by zh355245849 on 2017/5/25.
 */
public class ParkingLot {

    private final Level[] levels;

    public ParkingLot(int numOfLevels, int numSpotsPerLevel) {
        levels = new Level[numOfLevels];
        for (int i = 0; i < numOfLevels; i++) {
            levels[i] = new Level(numSpotsPerLevel);
        }
    }

    public boolean hasSpot(Vehicle v) {
        for (Level l : levels) {
            if (l.hasSpot(v)) {
                return true;
            }
        }
        return false;
    }

    public boolean park(Vehicle v) {
        for (Level l : levels) {
            if (l.park(v)) {
               return true;
            }
        }
        return false;
    }

    public boolean leave(Vehicle v) {
        for (Level l : levels) {
            if (l.leave(v)) {
                return true;
            }
        }
        return false;
    }
}
